+++ includes/class-newspack-newsletters-editor.php
@@ -53,7 +53,6 @@ final class Newspack_Newsletters_Editor {
 		add_action( 'rest_api_init', [ __CLASS__, 'add_newspack_author_info' ] );
 		add_filter( 'the_posts', [ __CLASS__, 'maybe_reset_excerpt_length' ] );
 		add_filter( 'should_load_remote_block_patterns', [ __CLASS__, 'strip_block_patterns' ] );
-		add_filter( 'enter_title_here', [ __CLASS__, 'placeholder_editor_title' ] );
 	}
 
 	/**
@@ -95,6 +94,34 @@ final class Newspack_Newsletters_Editor {
 	 */
 	private static function is_editing_email() {
 		return in_array( get_post_type(), self::get_email_editor_cpts() );
+	}
+
+	/**
+	 * Get CSS rules for color palette.
+	 *
+	 * @param string $container_selector Optional selector to prefix as a container to every rule.
+	 * 
+	 * @return string CSS rules.
+	 */
+	public static function get_color_palette_css( $container_selector = '' ) {
+		$rules = [];
+		// Add `.has-{color-name}-color` rules for each palette color.
+		$color_palette = json_decode( get_option( 'newspack_newsletters_color_palette', false ), true );
+		if ( ! empty( $color_palette ) ) {
+			foreach ( $color_palette as $color_name => $color_value ) {
+				$rules[] = '.has-' . esc_html( $color_name ) . '-color { color: ' . esc_html( $color_value ) . '; }';
+			}
+		}
+		if ( $container_selector ) {
+			$container_selector = esc_html( $container_selector );
+			$rules              = array_map(
+				function( $rule ) use ( $container_selector ) {
+					return $container_selector . ' ' . $rule;
+				},
+				$rules 
+			);
+		}
+		return implode( "\n", $rules );
 	}
 
 	/**
@@ -148,21 +175,6 @@ final class Newspack_Newsletters_Editor {
 		}
 
 		return $should_load_remote;
-	}
-
-	/**
-	 * Placeholder text for the editor title.
-	 *
-	 * @param string $title_placeholder Placeholder text for the title.
-	 *
-	 * @return string Placeholder text for the title.
-	 */
-	public static function placeholder_editor_title( $title_placeholder ) {
-		if ( self::is_editing_email() ) {
-			return __( 'Subject', 'newspack-newsletters' );
-		}
-
-		return $title_placeholder;
 	}
 
 	/**
@@ -221,6 +233,9 @@ final class Newspack_Newsletters_Editor {
 			'core/separator',
 			'core/list',
 			'core/quote',
+			'core/site-logo',
+			'core/site-tagline',
+			'core/site-title',
 			'core/social-links',
 			'core/social-link',
 			'newspack-newsletters/posts-inserter',
@@ -242,6 +257,8 @@ final class Newspack_Newsletters_Editor {
 			wp_style_add_data( 'newspack-newsletters', 'rtl', 'replace' );
 			wp_enqueue_style( 'newspack-newsletters' );
 
+			wp_add_inline_style( 'newspack-newsletters', self::get_color_palette_css( '.editor-styles-wrapper' ) );
+
 			\wp_enqueue_script(
 				'newspack-newsletters-editor',
 				plugins_url( '../dist/editor.js', __FILE__ ),
@@ -261,6 +278,8 @@ final class Newspack_Newsletters_Editor {
 					'mjml_handling_post_types' => $mjml_handling_post_types,
 				]
 			);
+
+			do_action( 'newspack_newsletters_enqueue_block_editor_assets' );
 		}
 
 		if ( self::is_editing_newsletter_ad() ) {