+++ includes/class-newspack-newsletters-renderer.php
@@ -47,12 +47,36 @@ final class Newspack_Newsletters_Renderer {
 	protected static $post_permalink = null;
 
 	/**
+	 * Inline tags that are allowed to be rendered in a text block.
+	 * 
+	 * @var bool[]|array[] Associative array of tag names to allowed attributes.
+	 */
+	public static $allowed_inline_tags = [
+		's'      => true,
+		'b'      => true,
+		'strong' => true,
+		'i'      => true,
+		'em'     => true,
+		'span'   => true,
+		'u'      => true,
+		'small'  => true,
+		'sub'    => true,
+		'sup'    => true,
+		'a'      => [
+			'href'   => true,
+			'target' => true,
+			'rel'    => true,
+		],
+	];
+
+	/**
 	 * Convert a list to HTML attributes.
 	 *
 	 * @param array $attributes Array of attributes.
 	 * @return string HTML attributes as a string.
 	 */
 	private static function array_to_attributes( $attributes ) {
+		$attributes = apply_filters( 'newspack_newsletters_mjml_component_attributes', $attributes );
 		return join(
 			' ',
 			array_map(
@@ -228,6 +252,26 @@ final class Newspack_Newsletters_Renderer {
 	}
 
 	/**
+	 * Whether the block is empty.
+	 *
+	 * @param WP_Block $block The block.
+	 *
+	 * @return bool Whether the block is empty.
+	 */
+	public static function is_empty_block( $block ) {
+		$blocks_without_inner_html = [
+			'core/site-logo',
+			'core/site-title',
+			'core/site-tagline',
+		];
+
+		$empty_block_name = empty( $block['blockName'] );
+		$empty_html       = ! in_array( $block['blockName'], $blocks_without_inner_html, true ) && empty( $block['innerHTML'] );
+
+		return $empty_block_name || $empty_html;
+	}
+
+	/**
 	 * Convert a Gutenberg block to an MJML component.
 	 * MJML component will be put in an mj-column in an mj-section for consistent layout,
 	 * unless it's a group or a columns block.
@@ -244,7 +288,7 @@ final class Newspack_Newsletters_Renderer {
 		$inner_blocks = $block['innerBlocks'];
 		$inner_html   = $block['innerHTML'];
 
-		if ( ! isset( $attrs['innerBlocksToInsert'] ) && ( empty( $block_name ) || empty( $inner_html ) ) ) {
+		if ( ! isset( $attrs['innerBlocksToInsert'] ) && self::is_empty_block( $block ) ) {
 			return '';
 		}
 
@@ -273,6 +317,11 @@ final class Newspack_Newsletters_Renderer {
 
 		$font_family = 'core/heading' === $block_name ? self::$font_header : self::$font_body;
 
+		if ( ! empty( $inner_html ) ) {
+			// Replace <mark /> with <span />.
+			$inner_html = preg_replace( '/<mark\s(.+?)>(.+?)<\/mark>/is', '<span $1>$2</span>', $inner_html );
+		}
+
 		switch ( $block_name ) {
 			/**
 			 * Text-based blocks.
@@ -281,6 +330,8 @@ final class Newspack_Newsletters_Renderer {
 			case 'core/list':
 			case 'core/heading':
 			case 'core/quote':
+			case 'core/site-title':
+			case 'core/site-tagline':
 			case 'newspack-newsletters/share':
 				$text_attrs = array_merge(
 					array(
@@ -297,6 +348,28 @@ final class Newspack_Newsletters_Renderer {
 					return '';
 				}
 
+				if ( 'core/site-tagline' === $block_name ) {
+					$inner_html = get_bloginfo( 'description' );
+				}
+
+				if ( 'core/site-title' === $block_name ) {
+					$inner_html = get_bloginfo( 'name' );
+					$tag_name   = 'h1';
+					if ( isset( $attrs['level'] ) ) {
+						$tag_name = 0 === $attrs['level'] ? 'p' : 'h' . (int) $attrs['level'];
+					}
+					if ( ! ( isset( $attrs['isLink'] ) && ! $attrs['isLink'] ) ) {
+						$link_attrs = array(
+							'href="' . esc_url( get_bloginfo( 'url' ) ) . '"',
+						);
+						if ( isset( $attrs['linkTarget'] ) && '_blank' === $attrs['linkTarget'] ) {
+							$link_attrs[] = 'target="_blank"';
+						}
+						$inner_html = sprintf( '<a %1$s>%2$s</a>', implode( ' ', $link_attrs ), esc_html( $inner_html ) );
+					}
+					$inner_html = sprintf( '<%1$s>%2$s</%1$s>', $tag_name, $inner_html );
+				}
+
 				// Only mj-text has to use container-background-color attr for background color.
 				if ( isset( $text_attrs['background-color'] ) ) {
 					$text_attrs['container-background-color'] = $text_attrs['background-color'];
@@ -307,17 +380,37 @@ final class Newspack_Newsletters_Renderer {
 				break;
 
 			/**
+			 * Site logo block.
+			 */
+			case 'core/site-logo':
+				$custom_logo_id = get_theme_mod( 'custom_logo' );
+				$image          = wp_get_attachment_image_src( $custom_logo_id, 'full' );
+				$markup         = '';
+				if ( ! empty( $image ) ) {
+					$img_attrs = array(
+						'padding' => '0',
+						'width'   => sprintf( '%spx', isset( $attrs['width'] ) ? $attrs['width'] : '125' ),
+						'align'   => isset( $attrs['align'] ) ? $attrs['align'] : 'left',
+						'src'     => $image[0],
+						'href'    => isset( $attrs['isLink'] ) && ! $attrs['isLink'] ? '' : esc_url( home_url( '/' ) ),
+						'target'  => isset( $attrs['linkTarget'] ) && '_blank' === $attrs['linkTarget'] ? '_blank' : '',
+					);
+					$markup   .= '<mj-image ' . self::array_to_attributes( $img_attrs ) . ' />';
+				}
+				$block_mjml_markup = $markup;
+				break;
+
+			/**
 			 * Image block.
 			 */
 			case 'core/image':
 				// Parse block content.
 				$dom = new DomDocument();
 				libxml_use_internal_errors( true );
-				$dom->loadHTML( mb_convert_encoding( $inner_html, 'HTML-ENTITIES', get_bloginfo( 'charset' ) ) );
-				$xpath      = new DOMXpath( $dom );
-				$img        = $xpath->query( '//img' )[0];
+				$dom->loadHTML( mb_convert_encoding( $inner_html, 'HTML-ENTITIES', get_bloginfo( 'charset' ) ), LIBXML_HTML_NOIMPLIED | LIBXML_HTML_NODEFDTD );
+				$img        = $dom->getElementsByTagName( 'img' )->item( 0 );
 				$img_src    = $img->getAttribute( 'src' );
-				$figcaption = $xpath->query( '//figcaption/text()' )[0];
+				$figcaption = $dom->getElementsByTagName( 'figcaption' )->item( 0 );
 
 				$img_attrs = array(
 					'padding' => '0',
@@ -360,15 +453,23 @@ final class Newspack_Newsletters_Renderer {
 				$markup = '<mj-image ' . self::array_to_attributes( $img_attrs ) . ' />';
 
 				if ( $figcaption ) {
+					$caption_html  = '';
+					$caption_nodes = $figcaption->childNodes; // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
+					foreach ( $caption_nodes as $caption_node ) {
+						$caption_html .= $dom->saveHTML( $caption_node );
+					}
 					$caption_attrs = array(
+						'css-class'   => 'image-caption',
 						'align'       => 'center',
 						'color'       => '#555d66',
 						'line-height' => '1.56',
 						'font-size'   => '13px',
 						'font-family' => $font_family,
 					);
-					 // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
-					$markup .= '<mj-text ' . self::array_to_attributes( $caption_attrs ) . '>' . $figcaption->wholeText . '</mj-text>';
+					$markup       .= '<mj-text ' . self::array_to_attributes( $caption_attrs ) . '>' . wp_kses(
+						$caption_html,
+						self::$allowed_inline_tags
+					) . '</mj-text>';
 				}
 
 				$block_mjml_markup = $markup;