+++ CHANGELOG.md
@@ -1,3 +1,143 @@
+# [1.44.0](https://github.com/Automattic/newspack-newsletters/compare/v1.43.0...v1.44.0) (2022-04-18)
+
+
+### Bug Fixes
+
+* **layouts:** ensure selected layout is saved ([#786](https://github.com/Automattic/newspack-newsletters/issues/786)) ([5c659d1](https://github.com/Automattic/newspack-newsletters/commit/5c659d1bc3952c6786d22b89b5022c58fad3777d))
+
+
+### Features
+
+* ActiveCampaign esp integration ([#775](https://github.com/Automattic/newspack-newsletters/issues/775)) ([0d38151](https://github.com/Automattic/newspack-newsletters/commit/0d381510dc86b8eb7c548c5e3c101dd1c3e6a42d))
+* add block support for Site Logo, Site Tagline, and Site Title ([#679](https://github.com/Automattic/newspack-newsletters/issues/679)) ([714b498](https://github.com/Automattic/newspack-newsletters/commit/714b4982b96690f031a9a0c509ebd1fd992d774e))
+* simplify visibility option ([#733](https://github.com/Automattic/newspack-newsletters/issues/733)) ([d7ea865](https://github.com/Automattic/newspack-newsletters/commit/d7ea86546ac0c62728362008669d65aa2e484dca))
+
+# [1.44.0-alpha.3](https://github.com/Automattic/newspack-newsletters/compare/v1.44.0-alpha.2...v1.44.0-alpha.3) (2022-04-13)
+
+
+### Bug Fixes
+
+* **layouts:** ensure selected layout is saved ([#786](https://github.com/Automattic/newspack-newsletters/issues/786)) ([5c659d1](https://github.com/Automattic/newspack-newsletters/commit/5c659d1bc3952c6786d22b89b5022c58fad3777d))
+
+# [1.44.0-alpha.2](https://github.com/Automattic/newspack-newsletters/compare/v1.44.0-alpha.1...v1.44.0-alpha.2) (2022-04-08)
+
+
+### Features
+
+* ActiveCampaign esp integration ([#775](https://github.com/Automattic/newspack-newsletters/issues/775)) ([0d38151](https://github.com/Automattic/newspack-newsletters/commit/0d381510dc86b8eb7c548c5e3c101dd1c3e6a42d))
+
+# [1.44.0-alpha.1](https://github.com/Automattic/newspack-newsletters/compare/v1.43.0...v1.44.0-alpha.1) (2022-04-07)
+
+
+### Features
+
+* add block support for Site Logo, Site Tagline, and Site Title ([#679](https://github.com/Automattic/newspack-newsletters/issues/679)) ([714b498](https://github.com/Automattic/newspack-newsletters/commit/714b4982b96690f031a9a0c509ebd1fd992d774e))
+* simplify visibility option ([#733](https://github.com/Automattic/newspack-newsletters/issues/733)) ([d7ea865](https://github.com/Automattic/newspack-newsletters/commit/d7ea86546ac0c62728362008669d65aa2e484dca))
+
+# [1.43.0](https://github.com/Automattic/newspack-newsletters/compare/v1.42.1...v1.43.0) (2022-04-05)
+
+
+### Bug Fixes
+
+* disable row group block variation ([#772](https://github.com/Automattic/newspack-newsletters/issues/772)) ([82224ff](https://github.com/Automattic/newspack-newsletters/commit/82224ff1e372ac6cc0e585f9fd774cf86e245e33))
+* move subject back to the sidebar ([#773](https://github.com/Automattic/newspack-newsletters/issues/773)) ([46aad75](https://github.com/Automattic/newspack-newsletters/commit/46aad752ba45f4b36ba9eaf7530a3a5952c57a2b))
+
+
+### Features
+
+* allow theme customization ([#766](https://github.com/Automattic/newspack-newsletters/issues/766)) ([ebf3cca](https://github.com/Automattic/newspack-newsletters/commit/ebf3ccaac14a0df5e8641ce03d9404141c939793))
+* minify and notify clipable email body ([#764](https://github.com/Automattic/newspack-newsletters/issues/764)) ([2c1c6a3](https://github.com/Automattic/newspack-newsletters/commit/2c1c6a32c3437b447324b4b33c26121668c5801e))
+
+# [1.43.0-alpha.1](https://github.com/Automattic/newspack-newsletters/compare/v1.42.1...v1.43.0-alpha.1) (2022-03-31)
+
+
+### Bug Fixes
+
+* disable row group block variation ([#772](https://github.com/Automattic/newspack-newsletters/issues/772)) ([82224ff](https://github.com/Automattic/newspack-newsletters/commit/82224ff1e372ac6cc0e585f9fd774cf86e245e33))
+* move subject back to the sidebar ([#773](https://github.com/Automattic/newspack-newsletters/issues/773)) ([46aad75](https://github.com/Automattic/newspack-newsletters/commit/46aad752ba45f4b36ba9eaf7530a3a5952c57a2b))
+
+
+### Features
+
+* allow theme customization ([#766](https://github.com/Automattic/newspack-newsletters/issues/766)) ([ebf3cca](https://github.com/Automattic/newspack-newsletters/commit/ebf3ccaac14a0df5e8641ce03d9404141c939793))
+* minify and notify clipable email body ([#764](https://github.com/Automattic/newspack-newsletters/issues/764)) ([2c1c6a3](https://github.com/Automattic/newspack-newsletters/commit/2c1c6a32c3437b447324b4b33c26121668c5801e))
+
+## [1.42.1](https://github.com/Automattic/newspack-newsletters/compare/v1.42.0...v1.42.1) (2022-03-22)
+
+
+### Bug Fixes
+
+* color palette css rules ([#757](https://github.com/Automattic/newspack-newsletters/issues/757)) ([2d929ac](https://github.com/Automattic/newspack-newsletters/commit/2d929ac8277b833aaceada281bd6eae73c465805))
+
+## [1.42.1-alpha.1](https://github.com/Automattic/newspack-newsletters/compare/v1.42.0...v1.42.1-alpha.1) (2022-03-15)
+
+
+### Bug Fixes
+
+* color palette css rules ([#757](https://github.com/Automattic/newspack-newsletters/issues/757)) ([2d929ac](https://github.com/Automattic/newspack-newsletters/commit/2d929ac8277b833aaceada281bd6eae73c465805))
+
+# [1.42.0](https://github.com/Automattic/newspack-newsletters/compare/v1.41.0...v1.42.0) (2022-03-08)
+
+
+### Bug Fixes
+
+* **mailchimp:** check send status ([#746](https://github.com/Automattic/newspack-newsletters/issues/746)) ([0be0ad9](https://github.com/Automattic/newspack-newsletters/commit/0be0ad9eafeedcfce4763002ec266313cbb0826a))
+* php warning when unregistering a block pattern that isn't registered ([#738](https://github.com/Automattic/newspack-newsletters/issues/738)) ([8a2d0f6](https://github.com/Automattic/newspack-newsletters/commit/8a2d0f601c18733df0068593d768d68b41fa4ed5))
+* support for image caption inner html ([#745](https://github.com/Automattic/newspack-newsletters/issues/745)) ([eed512c](https://github.com/Automattic/newspack-newsletters/commit/eed512c5144822ddd337ebe47af06b3ba081b48b))
+
+
+### Features
+
+* remove subject input (title) from the sidebar ([#735](https://github.com/Automattic/newspack-newsletters/issues/735)) ([969ff3b](https://github.com/Automattic/newspack-newsletters/commit/969ff3bfd1495d15333b1bf1c7d93287c39cffa5))
+
+# [1.42.0-alpha.1](https://github.com/Automattic/newspack-newsletters/compare/v1.41.0...v1.42.0-alpha.1) (2022-02-24)
+
+
+### Bug Fixes
+
+* **mailchimp:** check send status ([#746](https://github.com/Automattic/newspack-newsletters/issues/746)) ([0be0ad9](https://github.com/Automattic/newspack-newsletters/commit/0be0ad9eafeedcfce4763002ec266313cbb0826a))
+* php warning when unregistering a block pattern that isn't registered ([#738](https://github.com/Automattic/newspack-newsletters/issues/738)) ([8a2d0f6](https://github.com/Automattic/newspack-newsletters/commit/8a2d0f601c18733df0068593d768d68b41fa4ed5))
+* support for image caption inner html ([#745](https://github.com/Automattic/newspack-newsletters/issues/745)) ([eed512c](https://github.com/Automattic/newspack-newsletters/commit/eed512c5144822ddd337ebe47af06b3ba081b48b))
+
+
+### Features
+
+* remove subject input (title) from the sidebar ([#735](https://github.com/Automattic/newspack-newsletters/issues/735)) ([969ff3b](https://github.com/Automattic/newspack-newsletters/commit/969ff3bfd1495d15333b1bf1c7d93287c39cffa5))
+
+# [1.41.0](https://github.com/Automattic/newspack-newsletters/compare/v1.40.0...v1.41.0) (2022-02-22)
+
+
+### Bug Fixes
+
+* font size vertical rhythm for quote and paragraph ([#725](https://github.com/Automattic/newspack-newsletters/issues/725)) ([d656902](https://github.com/Automattic/newspack-newsletters/commit/d65690250d8e7d6a133cf834cf3e73500dfd109b))
+* iframe styles for post inserter ([#723](https://github.com/Automattic/newspack-newsletters/issues/723)) ([d448daf](https://github.com/Automattic/newspack-newsletters/commit/d448daff469df1573496be617e215bbc0d6bf8fd))
+* remove duotone filters ([#732](https://github.com/Automattic/newspack-newsletters/issues/732)) ([b1763c0](https://github.com/Automattic/newspack-newsletters/commit/b1763c097962596ffbc6d96edf04c99c30c8f47b))
+
+
+### Features
+
+* **post-inserter:** add featured image size option ([#712](https://github.com/Automattic/newspack-newsletters/issues/712)) ([b20b132](https://github.com/Automattic/newspack-newsletters/commit/b20b132868b593e5e54646ead212cfd51ebf6e67))
+* remove wp-bundled block patterns ([#731](https://github.com/Automattic/newspack-newsletters/issues/731)) ([de0be75](https://github.com/Automattic/newspack-newsletters/commit/de0be75dc4d059cf1721c981903b4b43a9975101))
+* support block padding ([#724](https://github.com/Automattic/newspack-newsletters/issues/724)) ([877f07f](https://github.com/Automattic/newspack-newsletters/commit/877f07fa52ad06247d15bbde2c2e99ceb7fdc623))
+* update init modal design ([#710](https://github.com/Automattic/newspack-newsletters/issues/710)) ([5c8e273](https://github.com/Automattic/newspack-newsletters/commit/5c8e273a275c4fb5bd9e93a8b899ebfd5761aff7))
+
+# [1.41.0-alpha.1](https://github.com/Automattic/newspack-newsletters/compare/v1.40.0...v1.41.0-alpha.1) (2022-02-10)
+
+
+### Bug Fixes
+
+* font size vertical rhythm for quote and paragraph ([#725](https://github.com/Automattic/newspack-newsletters/issues/725)) ([d656902](https://github.com/Automattic/newspack-newsletters/commit/d65690250d8e7d6a133cf834cf3e73500dfd109b))
+* iframe styles for post inserter ([#723](https://github.com/Automattic/newspack-newsletters/issues/723)) ([d448daf](https://github.com/Automattic/newspack-newsletters/commit/d448daff469df1573496be617e215bbc0d6bf8fd))
+* remove duotone filters ([#732](https://github.com/Automattic/newspack-newsletters/issues/732)) ([b1763c0](https://github.com/Automattic/newspack-newsletters/commit/b1763c097962596ffbc6d96edf04c99c30c8f47b))
+
+
+### Features
+
+* **post-inserter:** add featured image size option ([#712](https://github.com/Automattic/newspack-newsletters/issues/712)) ([b20b132](https://github.com/Automattic/newspack-newsletters/commit/b20b132868b593e5e54646ead212cfd51ebf6e67))
+* remove wp-bundled block patterns ([#731](https://github.com/Automattic/newspack-newsletters/issues/731)) ([de0be75](https://github.com/Automattic/newspack-newsletters/commit/de0be75dc4d059cf1721c981903b4b43a9975101))
+* support block padding ([#724](https://github.com/Automattic/newspack-newsletters/issues/724)) ([877f07f](https://github.com/Automattic/newspack-newsletters/commit/877f07fa52ad06247d15bbde2c2e99ceb7fdc623))
+* update init modal design ([#710](https://github.com/Automattic/newspack-newsletters/issues/710)) ([5c8e273](https://github.com/Automattic/newspack-newsletters/commit/5c8e273a275c4fb5bd9e93a8b899ebfd5761aff7))
+
 # [1.40.0](https://github.com/Automattic/newspack-newsletters/compare/v1.39.0...v1.40.0) (2022-02-08)